+++ pkg/hintrunner/operand_test.go
@@ -6,7 +6,6 @@ import (
 
 	VM "github.com/NethermindEth/cairo-vm-go/pkg/vm"
 	"github.com/NethermindEth/cairo-vm-go/pkg/vm/memory"
-	f "github.com/consensys/gnark-crypto/ecc/stark-curve/fp"
 	"github.com/stretchr/testify/require"
 )
 
@@ -180,18 +179,4 @@ func TestResolveMulOp(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, memory.MemoryValueFromInt(500), res)
 
-}
-
-func defaultVirtualMachine() *VM.VirtualMachine {
-	vm, _ := VM.NewVirtualMachine(make([]*f.Element, 0), VM.VirtualMachineConfig{})
-	return vm
-}
-
-func writeTo(vm *VM.VirtualMachine, segment uint64, offset uint64, val *memory.MemoryValue) {
-	_ = vm.MemoryManager.Memory.Write(segment, offset, val)
-}
-
-func readFrom(vm *VM.VirtualMachine, segment uint64, offset uint64) *memory.MemoryValue {
-	val, _ := vm.MemoryManager.Memory.Read(segment, offset)
-	return val
 }
